refactor(support): 패키징 변경
