+++ app/src/main/java/woowacourse/shopping/support/framework/presentation/IntentDataNullErrorProcsser.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.util
+package woowacourse.shopping.support.framework.presentation
 
 import android.app.Activity
 import android.util.Log
