+++ app/src/main/java/woowacourse/shopping/support/framework/presentation/RecyclerViewExtension.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.util
+package woowacourse.shopping.support.framework.presentation
 
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.SimpleItemAnimator
