+++ app/src/main/java/woowacourse/shopping/ui/serversetting/ServerSettingActivity.kt
@@ -6,8 +6,8 @@ import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.data.httpclient.RetrofitModule
+import woowacourse.shopping.support.framework.presentation.setThrottleFirstOnClickListener
 import woowacourse.shopping.ui.shopping.ShoppingActivity
-import woowacourse.shopping.util.setThrottleFirstOnClickListener
 
 class ServerSettingActivity : AppCompatActivity() {
