+++ app/src/main/java/woowacourse/shopping/ui/shopping/recentproduct/RecentProductViewHolder.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ItemRecentProductBinding
+import woowacourse.shopping.support.framework.presentation.setThrottleFirstOnClickListener
 import woowacourse.shopping.ui.model.UiRecentProduct
-import woowacourse.shopping.util.setThrottleFirstOnClickListener
 
 class RecentProductViewHolder(parent: ViewGroup, onItemClick: (Int) -> Unit) :
     RecyclerView.ViewHolder(
